refactor main.py for more clarity
